Create I2Pservice as a way to integrate service management, hide unnecessary handlers
