+++ I2PTunnel.h
@@ -9,6 +9,7 @@
 #include "Identity.h"
 #include "Destination.h"
 #include "Streaming.h"
+#include "I2PService.h"
 
 namespace i2p
 {
@@ -17,21 +18,19 @@ namespace client
 	const size_t I2P_TUNNEL_CONNECTION_BUFFER_SIZE = 8192;
 	const int I2P_TUNNEL_CONNECTION_MAX_IDLE = 3600; // in seconds	
 	const int I2P_TUNNEL_DESTINATION_REQUEST_TIMEOUT = 10; // in seconds
-	const i2p::data::SigningKeyType I2P_TUNNEL_DEFAULT_KEY_TYPE = i2p::data::SIGNING_KEY_TYPE_ECDSA_SHA256_P256;
 
-	class I2PTunnel;
-	class I2PTunnelConnection: public std::enable_shared_from_this<I2PTunnelConnection>
+
+	class I2PTunnelConnection: public I2PServiceHandler, public std::enable_shared_from_this<I2PTunnelConnection>
 	{
 		public:
 
-			I2PTunnelConnection (I2PTunnel * owner, boost::asio::ip::tcp::socket * socket,
+			I2PTunnelConnection (I2PService * owner, boost::asio::ip::tcp::socket * socket,
 				const i2p::data::LeaseSet * leaseSet); // to I2P
-			I2PTunnelConnection (I2PTunnel * owner, boost::asio::ip::tcp::socket * socket,
+			I2PTunnelConnection (I2PService * owner, boost::asio::ip::tcp::socket * socket,
 				std::shared_ptr<i2p::stream::Stream> stream); // to I2P using simplified API :)
-			I2PTunnelConnection (I2PTunnel * owner, std::shared_ptr<i2p::stream::Stream> stream,  boost::asio::ip::tcp::socket * socket, 
+			I2PTunnelConnection (I2PService * owner, std::shared_ptr<i2p::stream::Stream> stream,  boost::asio::ip::tcp::socket * socket, 
 				const boost::asio::ip::tcp::endpoint& target, bool quiet = true); // from I2P
 			~I2PTunnelConnection ();
-
 			void I2PConnect (const uint8_t * msg = nullptr, size_t len = 0);
 			void Connect ();
 			
@@ -52,33 +51,11 @@ namespace client
 			uint8_t m_Buffer[I2P_TUNNEL_CONNECTION_BUFFER_SIZE], m_StreamBuffer[I2P_TUNNEL_CONNECTION_BUFFER_SIZE];
 			boost::asio::ip::tcp::socket * m_Socket;
 			std::shared_ptr<i2p::stream::Stream> m_Stream;
-			I2PTunnel * m_Owner;
 			boost::asio::ip::tcp::endpoint m_RemoteEndpoint;
 			bool m_IsQuiet; // don't send destination
-	};	
+	};
 
-	class I2PTunnel
-	{
-		public:
-
-			I2PTunnel (ClientDestination * localDestination  = nullptr);
-			virtual ~I2PTunnel () { ClearConnections (); }; 
-
-			void AddConnection (std::shared_ptr<I2PTunnelConnection> conn);
-			void RemoveConnection (std::shared_ptr<I2PTunnelConnection> conn);	
-			void ClearConnections ();
-			ClientDestination * GetLocalDestination () { return m_LocalDestination; };
-			void SetLocalDestination (ClientDestination * dest) { m_LocalDestination = dest; }; 			
-
-			boost::asio::io_service& GetService () { return m_LocalDestination->GetService (); };
-			
-		private:
-
-			ClientDestination * m_LocalDestination;
-			std::set<std::shared_ptr<I2PTunnelConnection> > m_Connections;
-	};	
-	
-	class I2PClientTunnel: public I2PTunnel
+	class I2PClientTunnel: public I2PService
 	{
 		public:
 
@@ -103,7 +80,7 @@ namespace client
 			const i2p::data::IdentHash * m_DestinationIdentHash;
 	};	
 
-	class I2PServerTunnel: public I2PTunnel
+	class I2PServerTunnel: public I2PService
 	{
 		public:
 
@@ -121,7 +98,7 @@ namespace client
 
 			boost::asio::ip::tcp::endpoint m_Endpoint;		
 	};
-}		
+}
 }	
 
 #endif
